feat: add real and projected Community Value Index to the community stats draft
The weekly community-stats draft generator now reports both Community
Value Index figures alongside the SocketRelay, Directory, and
ServiceCredits aggregates: the real index (value already exchanged) and
the projected "value waiting to happen" figure (what open posts would add
if they all closed). Both are labeled as index numbers, not money, and the
drafting prompt tells the writer model the same so the framing survives
into the Quora draft.

To avoid a third hand-copied set of queries, the recognition source SQL
and contribution weights moved from scripts/recognizeGdp.mjs into a shared
module scripts/lib/gdpValueIndex.mjs, which also carries a script-side
mirror of the projection sources from packages/web/lib/gdp/projection.ts.
recognizeGdp.mjs imports from it now; its queries, weights, and
snapshot-write behavior are unchanged.

Operational scripts only — no app code, schema, route, or contract change.
