+++ .github/workflows/README.md
@@ -56,7 +56,7 @@ deploy's status (live or failed) — it does not build or deploy anything itself
 | `expo-android-scheduled-build.yml` | Expo — Scheduled Android Test Build | Mon/Wed/Fri 08:41 UTC; manual | Hands-free `preview` Android test build. Checks the free-tier quota first (skips when this month's automated budget is used) so the 15-builds/month allotment is never overspent; if the build fails, a Claude agent opens a fix branch and PR. |
 | `expo-preview.yml` | Expo — Preview APK | Mobile PRs labeled `build-apk`; manual | Builds a preview Android APK for a mobile PR (opt-in via label to save free-tier builds) and posts an install link. |
 | `expo-update.yml` | Expo — OTA Update | Push to `main` touching mobile code | Publishes a JavaScript/asset-only over-the-air update via EAS Update — no native rebuild; installed apps pick it up on next launch. |
-| `generate-community-stats.yml` | Generate — Community Stats Draft | Mondays 14:00 UTC; manual | Reads privacy-safe whole-community counts (open SocketRelay posts, Directory profiles and skills), drafts a Quora post with Claude, and files it as an issue for review. |
+| `generate-community-stats.yml` | Generate — Community Stats Draft | Mondays 14:00 UTC; manual | Reads privacy-safe whole-community counts (open SocketRelay posts, Directory profiles and skills, ServiceCredits aggregates, and the real + projected Community Value Index figures), drafts a Quora post with Claude, and files it as an issue for review. |
 | `generate-product-update.yml` | Generate — Product Update | Weekly (Thursdays 00:33 UTC); manual | Turns recent feat/fix/perf commits into a product update: publishes to the wiki, the blog registry, and the in-app feed, files a Quora draft issue, and tags the update. |
 | `generate-user-guide.yml` | User Guide — Regenerate | Mondays 06:00 UTC; manual | Regenerates the public user guide (`/guide` + `docs/USER_GUIDE.md`) from each plugin's inventory User Features section and test-script Core smoke steps, via `ctf/scripts/generate-user-guide.mjs`. |
 | `github-actions-billing-token-reminder.yml` | GitHub Actions — Billing Token Rotation Reminder | Mondays 12:00 UTC; manual | Files a reminder issue every ~45 days to rotate the Actions billing token, with a runbook checklist. |
