+++ ctf/docs/developer/test-scripts/gdp-test-script.md
@@ -402,6 +402,14 @@ The member count shown on the GDP dashboard must equal the member count shown in
 
 Pulled from §9 of the feature inventory:
 
+0. **The weekly community-stats draft reports both index figures, but is not member-testable here.**
+   The Monday community-stats draft issue (label `community-stats`, generated by
+   `ctf/scripts/generate-community-stats.mjs`) includes the real Community Value Index and the
+   projected "value waiting to happen" figure, computed from the same shared source SQL and weights
+   the dashboard uses (`ctf/scripts/lib/gdpValueIndex.mjs`). It runs from a scheduled GitHub Action
+   against the production database, so there is no in-app surface to test; sanity-check it by
+   comparing the two numbers in the next draft issue against the `/apps/gdp` dashboard on the same
+   day (they should match to rounding). Not a member-facing test.
 1. **Metric ownership roster not surfaced.** Ownership assignments for economics metrics are documented in contracts but there is no single roster page. A missing owner page is not a bug.
 2. **No plugin-specific cross-region transfer contract.** Regional/legal constraints for authenticated cross-region GDP publication are governed by platform defaults; a GDP-specific transfer-control contract has not been finalized. The absence of a plugin-level override is not a bug.
 3. **No explicit SLA document for snapshot publication.** Snapshot publication SLA and freeze windows follow operational best-effort; a formal SLA document has not been published. Missing SLA documentation is not a testable bug here.
